requirements.txt
Flask==0.10.1
dota2py==0.1.3
numpy==1.10.1
progressbar2
pymongo==2.6.3
scikit-learn==0.14.1
scipy==0.16.1
